Add a workaround in boot.sh for noexec & nosuid /run
WSL mounts /run with the nosuid and noexec mount flags. By doing so, it
prevents /run/setuid-programs from fulfilling its purpose as a place to
store executables with the setuid bits set.

This change adds a workaround to amend this issue. In the boot.sh
script, a few commands are added to transparently use
/var/run/setuid-programs instead. These added commands are performed
after the invocation of $GUIX_NEW_SYSTEM/boot (which initially populates
the setuid programs) and do the following:

 - remove the previous /var/run/setuid-programs, if there was one; this
   may be over-cautious, but ensures that the setuid programs created by
   the prior guix bootup are exactly the ones that are visible, without
   any additional inclusions
 - copy the binaries from /run/setuid-programs to /var/run/setuid-programs
 - bind-mount /var/run/setuid-programs onto /run/setuid-programs

Signed-off-by: Mike Petersen <mapsmap55@gmail.com>